refactor: Move SearchFilterSettingsModule to modules dir
